backend/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
brevitas==0.10.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==15.0.1
concrete-ml==1.9.0
concrete-ml-extensions==0.1.9
concrete-python==2.10.0
contourpy==1.3.0
cycler==0.12.1
dependencies==2.0.1
dill==0.4.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hummingbird-ml==0.4.11
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonpickle==4.1.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
onnx==1.17.0
onnxconverter-common==1.16.0
onnxmltools==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.18.0
packaging==25.0
pandas==2.0.3
pillow==11.3.0
protobuf==3.20.3
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.17.0
skl2onnx==1.19.1
skops==0.5.0
skorch==0.11.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.3.1
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unfoldNd==0.2.3
urllib3==2.5.0
uvicorn==0.35.0
xgboost==1.6.2
z3-solver==4.13.0.0
zipp==3.23.0
